+++ lib/benchee/formatters/console.ex
@@ -20,31 +20,67 @@ defmodule Benchee.Formatters.Console do
   def output(suite) do
     suite
     |> format
-    |> IO.puts
+    |> IO.write
   end
 
   @doc """
   Formats the benchmark statistics to a report suitable for output on the CLI.
 
+  Returns a list of lists, where each list element is a group belonging to one
+  specific input. So if there only was one (or no) input given through `:inputs`
+  then there's just one list inside.
+
   ## Examples
 
   ```
   iex> jobs = %{ "My Job" =>%{average: 200.0, ips: 5000.0,std_dev_ratio: 0.1, median: 190.0}, "Job 2" => %{average: 400.0, ips: 2500.0, std_dev_ratio: 0.2, median: 390.0}}
-  iex> Benchee.Formatters.Console.format(%{statistics: jobs, config: %{console: %{comparison: false, unit_scaling: :best}}})
-  ["\nName             ips        average  deviation         median\n",
+  iex> inputs = %{"My input" => jobs}
+  iex> Benchee.Formatters.Console.format(%{statistics: inputs, config: %{console: %{comparison: false, unit_scaling: :best}}})
+  [["\n##### With input My input #####", "\nName             ips        average  deviation         median\n",
   "My Job        5.00 K      200.00 μs    ±10.00%      190.00 μs\n",
-  "Job 2         2.50 K      400.00 μs    ±20.00%      390.00 μs"]
+  "Job 2         2.50 K      400.00 μs    ±20.00%      390.00 μs\n"]]
 
   ```
 
   """
-  def format(%{statistics: job_stats, config: %{console: config}}) do
+  def format(%{statistics: jobs_per_input, config: %{console: config}}) do
+    jobs_per_input
+    |> Enum.map(fn({input, jobs_stats}) ->
+         [input_header(input) | format_jobs(jobs_stats, config)]
+       end)
+  end
+
+  defp input_header(input) do
+    no_input_marker = Benchee.Benchmark.no_input
+    case input do
+      ^no_input_marker -> ""
+      _                -> "\n##### With input #{input} #####"
+    end
+  end
+
+  @doc """
+  Formats the job statistics to a report suitable for output on the CLI.
+
+  ## Examples
+
+  ```
+  iex> jobs = %{ "My Job" =>%{average: 200.0, ips: 5000.0,std_dev_ratio: 0.1, median: 190.0}, "Job 2" => %{average: 400.0, ips: 2500.0, std_dev_ratio: 0.2, median: 390.0}}
+  iex> Benchee.Formatters.Console.format_jobs(jobs, %{comparison: false, unit_scaling: :best})
+  ["\nName             ips        average  deviation         median\n",
+  "My Job        5.00 K      200.00 μs    ±10.00%      190.00 μs\n",
+  "Job 2         2.50 K      400.00 μs    ±20.00%      390.00 μs\n"]
+
+  ```
+
+  """
+  def format_jobs(job_stats, config) do
     sorted_stats = Statistics.sort(job_stats)
     units = units(sorted_stats, config)
     label_width = label_width job_stats
-    [column_descriptors(label_width) | job_reports(sorted_stats, units, label_width)
+
+    [column_descriptors(label_width) |
+      job_reports(sorted_stats, units, label_width)
       ++ comparison_report(sorted_stats, units, label_width, config)]
-    |> remove_last_blank_line
   end
 
   defp column_descriptors(label_width) do
@@ -64,7 +100,7 @@ defmodule Benchee.Formatters.Console do
   end
 
   defp job_reports(jobs, units, label_width) do
-    Enum.map(jobs, fn(job) -> format_job job, units, label_width end)
+    Enum.map(jobs, fn(job) -> format_jobs job, units, label_width end)
   end
 
   defp units(jobs, %{unit_scaling: scaling_strategy}) do
@@ -88,7 +124,7 @@ defmodule Benchee.Formatters.Console do
     }
   end
 
-  defp format_job({name, %{average:       average,
+  defp format_jobs({name, %{average:       average,
                            ips:           ips,
                            std_dev_ratio: std_dev_ratio,
                            median:        median}
@@ -151,12 +187,4 @@ defmodule Benchee.Formatters.Console do
   defp comparison_descriptor do
     "\nComparison: \n"
   end
-
-  defp remove_last_blank_line([head]) do
-    [String.rstrip(head)]
-  end
-  defp remove_last_blank_line([head | tail]) do
-    [head | remove_last_blank_line(tail)]
-  end
-
 end