+++ lib/benchee/statistics.ex
@@ -5,6 +5,7 @@ defmodule Benchee.Statistics do
   """
 
   alias Benchee.{Statistics, Conversion.Duration}
+  import Benchee.Utility.MapValues
   require Integer
 
   @doc """
@@ -36,33 +37,35 @@ defmodule Benchee.Statistics do
   ## Examples
 
       iex> run_times = [200, 400, 400, 400, 500, 500, 700, 900]
-      iex> suite = %{run_times: %{"My Job" => run_times}}
+      iex> suite = %{run_times: %{"Input" => %{"My Job" => run_times}}}
       iex> Benchee.Statistics.statistics(suite)
       %{
         statistics: %{
-          "My Job" => %{
-            average:       500.0,
-            ips:           2000.0,
-            std_dev:       200.0,
-            std_dev_ratio: 0.4,
-            std_dev_ips:   800.0,
-            median:        450.0,
-            minimum:       200,
-            maximum:       900,
-            sample_size:   8
+          "Input" => %{
+            "My Job" => %{
+              average:       500.0,
+              ips:           2000.0,
+              std_dev:       200.0,
+              std_dev_ratio: 0.4,
+              std_dev_ips:   800.0,
+              median:        450.0,
+              minimum:       200,
+              maximum:       900,
+              sample_size:   8
+            }
           }
         },
-        run_times: %{"My Job" => [200, 400, 400, 400, 500, 500, 700, 900]}
+        run_times: %{
+          "Input" => %{
+            "My Job" => [200, 400, 400, 400, 500, 500, 700, 900]
+          }
+        }
       }
 
   """
-  def statistics(suite = %{run_times: run_times}) do
-    statistics =
-      run_times
-      |> Enum.map(fn({name, job_run_times}) ->
-          {name, Statistics.job_statistics(job_run_times)}
-        end)
-      |> Map.new
+  def statistics(suite = %{run_times: run_times_per_input}) do
+    statistics = run_times_per_input
+                 |> map_values(&Statistics.job_statistics/1)
 
     Map.put suite, :statistics, statistics
   end