+++ test/benchee/utility/map_value_test.exs
@@ -1,0 +1,4 @@
+defmodule Benchee.Utility.MapValuesTest do
+  use ExUnit.Case, async: true
+  doctest Benchee.Utility.MapValues
+end